Fix `Connection::next`.
The current implementation based on `futures::select!` may miss values
produced by some of the selected futures. Because only one value is
(pseudo-randomly) returned when multiple futures are ready, the
non-selected one may be ignored. This can happen for instance if we
got a new inbound connection stream, which we return to the caller which
causes all futures to be dropped and the other values that were ready to
be dropped too. We need to make sure to always process all produced
values which is what this PR attempts to do.