Add `Drop` impl for `Connection`.
We want to make sure that even if users do not drive
`Connection::next_stream` to completion that all `Stream`s are closed
and pending `Waker`s notified when we have reached end of life, so we
close them all in a `Drop` impl.

Also, since we close the socket anyway only when things work out the way
they should be we move the close call to the location where it is
expected to happen, namely when reaching the end of the control command
queue.