Add `Connection::is_closed` flag.
This PR adds an `is_closed` flag to `Connection` and if `true`,
`Connection::next_stream` returns immediately with `Ok(None)` to help
clients that keep calling this method after they received an error.

While the API contract of `Connection::next_stream` already states that
after `Ok(None)` or `Err(_)` have been returned no further invocations
of `next_stream` must happen, it is conceivable that clients still do
this (e.g. trying to go on after `Err(_)`) so it seems advisable to
handle this case in a more explicit way.