If `lazy_open` use the default receive window.
The current implementation always uses the configured receive window
for the remote. However, when not sending the initial window update,
this means that the remote stops sending before having used up its
credit which it assumes matches the default receive window as it has
not received an initial window update. Therefore, if `lazy_open` is
true, the default receive window must be used initially.